+++ json_serializable/lib/src/field_helpers.dart
@@ -58,9 +58,8 @@ class _FieldSet implements Comparable<_FieldSet> {
     /// Returns the offset of given field/property in its source file – with a
     /// preference for the getter if it's defined.
     int _offsetFor(FieldElement e) {
-      if (e.getter != null && e.getter!.nameOffset != e.nameOffset) {
-        assert(e.nameOffset == -1);
-        return e.getter!.nameOffset;
+      if (e.isSynthetic) {
+        return (e.getter ?? e.setter)!.nameOffset;
       }
       return e.nameOffset;
     }